+++ app/src/main/java/s2017s40/kr/hs/mirim/remember_hi/DTO/UserDTO.java
@@ -5,12 +5,14 @@ public class UserDTO {
     private String email;
     private int age;
     private  String gender;
+    String phoneNum;
     UserDTO(){ }
-    UserDTO(String name, String email, int age, String gender){
+    UserDTO(String name, String email, int age, String gender, String phoneNum){
         this.age = age;
         this.email = email;
         this.gender = gender;
         this.name = name;
+        this.phoneNum = phoneNum;
     }
 
     public void setGender(String gender) {
