+++ app/src/main/java/s2017s40/kr/hs/mirim/remember_hi/SignUpActivity.java
@@ -1,0 +1,13 @@
+package s2017s40.kr.hs.mirim.remember_hi;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class SignUpActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_sign_up);
+    }
+}
